+++ ACG/Scenegraph/CoordFrameNode.cc
@@ -94,7 +94,7 @@ CoordFrameNode::availableDrawModes() const
 
 
 void 
-CoordFrameNode::boundingBox(Vec3f& _bbMin, Vec3f& _bbMax)
+CoordFrameNode::boundingBox(Vec3d& _bbMin, Vec3d& _bbMax)
 {
   _bbMin.minimize(bb_min_);
   _bbMax.maximize(bb_max_);
