+++ ACG/Scenegraph/PlaneNode.hh
@@ -89,7 +89,7 @@ class ACGDLLEXPORT PlaneNode : public BaseNode
     unsigned int availableDrawModes() const;
 
     /// update bounding box
-    void boundingBox(Vec3f& _bbMin, Vec3f& _bbMax);
+    void boundingBox(Vec3d& _bbMin, Vec3d& _bbMax);
 
     /// draw Plane
     void draw(GLState& _state, unsigned int _drawMode);
