+++ ACG/Scenegraph/StatusNodesT.hh
@@ -103,7 +103,7 @@ public:
 
 
   unsigned int availableDrawModes() const;
-  void boundingBox(Vec3f& _bbMin, Vec3f& _bbMax);
+  void boundingBox(Vec3d& _bbMin, Vec3d& _bbMax);
   void draw(GLState& _state, unsigned int _drawMode);
   void pick(GLState& /* _state */ , PickTarget /* _target */ ) {}
 
@@ -128,8 +128,8 @@ private:
   std::vector<FaceHandle>    fh_cache_;
 
   // bounding box
-  Vec3f bbMin_;
-  Vec3f bbMax_;
+  Vec3d bbMin_;
+  Vec3d bbMax_;
 };
